Add OTFI parser & injector to the repository.
Signed-off-by: Pascal Nasahl <nasahl@google.com>

Update readme

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add python requirements file and update readme.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Allow to specify the output name of the pickle graph.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Update instructions for the yosys flow in the README.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Store pin information into the edges of the graph.

The edges of the graph now includes information about the pin name (e.g., Q, D) for each input and output pin.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add first part of the fi_injector tool.

This tool extracts the target graph containing the fault sensitive parts
of the circuit. This target graph and the list of all possible fault
location combinations then is used to induce the faults.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Rename nodes of the extracted graphs.

Renaming is necessary to break the dependencies between the extracted
graphs before merging them into a single graph.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Set in- and outnodes for the extracted graph.

For each extracted subgraph, the type of the input and output node is
set. For circular graphs, split the common in- and outnode to two nodes.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add missing input nodes to the extracted graph.

The extracted target graph is a subview of the original graph. In this
subview, some input nodes for the nodes are missing and needs to be
added.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Connect the subgraphs of the target graph.

The extracted target graph consists of several subgraphs with an input
and output node each. Connect these subgraphs by connecting the outputs
with the inputs.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add injector class to the repo.

The injector class is responsible for creating the faulty and
differential graph. The differential graph then is converted to a
boolean formula, which can be handed to a SAT solver.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Create the differential graph.

The differential graph is created by inducing faults into the faulty
target graph and by merging this graph with the non-faulty copy. Input
and output values are set according to the fault model file.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add ports and pins to the nodes.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Remove exit.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Use ray to distribute FI insertion.

Ray is used to distribute the insertion of faults in a certain fault
location in parallel.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Rename ports of the output logic AND.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Node handling.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Remove exit

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Remap input pins for faulty nodes.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Fix string comparison bug.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Fix the handling of input and output pin names.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add ray to requirements.txt

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Address PR reviews.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Apply linter.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Private methods.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add type hinting for methods.

Signed-off-by: Pascal Nasahl <nasahl@google.com>