Add the formula class.
The FormulaBuilder and the NANGATE cell_library convert a circuit graph
into a boolean formula in CNF using the Tseitin transformation.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Set certain variables to a predefined value.

Allows to set clk, rst, and null/one nodes to a predefined value.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Create CNF and add output AND.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add new gate types for the cell library.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Apply linter to the cell library.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add the handling of in and out nodes.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add mapping for input pins.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Create the boolean formula with the given cell lib.

The boolean formula in CNF of the differential graph is created using
the Tseitin transformation. This includes the handling of the input and
output logic.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add license header.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add type annotation for functions.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Address PR reviews.

The cell library is now renamed, a data class is used for the input
pins, and some other minor fixes are applied.

Signed-off-by: Pascal Nasahl <nasahl@google.com>