Enable GitHub actions for CI.
Add GitHub actions configuration file and adapt code to work with pytest.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Change default branch for testing.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Establish testing for the parser and the injector.

For CI, the parser and the injector are extended to support pytest. For
this, a small fault model is added and the argument parser is changed.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Change branches for GitHub actions.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Add timeout for CI job.

Signed-off-by: Pascal Nasahl <nasahl@google.com>