Add the cell library generator script.
The cell library generator is responsible of parsing a cell library in
the liberty format and convert it to the format needed by OTFI. Here,
the most important step is the convertion of the cell formulas to
clauses, which can be processed by the SAT solver.

Signed-off-by: Pascal Nasahl <nasahl@google.com>

Convert formula string to sympy.

Signed-off-by: Pascal Nasahl <nasahl@google.com>