Autogenerate fault locations.
The "auto_fl" command line argument now allows to automatically create the
fault locations. For this, all comb. gates of the target circuit are
extracted and are used for the FI injection.

Hence, there is now no need to manually define the fault locations in
the fault_model.json file.

Signed-off-by: Pascal Nasahl <nasahl@google.com>