Add the cell library generator script.
The cell library generator is responsible of parsing a cell library in
the liberty format and convert it to the format needed by OTFI. Here,
the most important step is the convertion of the cell formulas to
clauses, which can be processed by the SAT solver. Mako and the template
file is used to generate the cell_lib.py in the end.

Signed-off-by: Pascal Nasahl <nasahl@google.com>