Run multi-bit experiments and adapt the framework.
In this PR, multi-bit inputs and outputs were tested. For this, a new
example fault model file was created and the code was fixed with the
following changes:
- Adapted the output logic. To enable multi-bit inputs/outputs, all
output signals for the faulty circuit are connected with a logical OR
and the output is connected to the final logical AND. This sets the
output clause to 1, when a output signal changes. For the new output
logic, the logical OR also was added to the OTFI specific part of the
cell lib.
-Adapted the FI Injector to handle multiple inputs/outputs for each
stage.
-Fixed a bug in the input validation. The renaming of
certain pins of a cell (e.g., I and S) is necessary, as sympy already
uses these names.

The README in the example folder was updated to explain the multi-bit
experiment.

Signed-off-by: Pascal Nasahl <nasahl@google.com>