Add possibility to define expected fault outputs.
When verifying state machines, we want to have the possibility to check,
if we could reach a certain state. In this PR, a new feature allowing to
define expected fault values is added. The framework now tries to find a
fault which changes the output of the circuit to the expected fault
value.

Signed-off-by: Pascal Nasahl <nasahl@google.com>