Limit number of fault locations.
As the possible fault combinations is exponentially increasing with the
number of simultaneous faults, the workload gets unhandleable for a
large number of sim. faults. This PR adds the -l argument to the FI
injector allowing to set a max. number of fault locations. This allows
us to probabilistically test the circuit.

Signed-off-by: Pascal Nasahl <nasahl@google.com>